more generic op handling
